smyrna warning squashing

See merge request graphviz/graphviz!2830

+++ cmd/smyrna/gvprpipe.h
@@ -1,5 +1,5 @@
 /*************************************************************************
- * Copyright (c) 2011 AT&T Intellectual Property 
+ * Copyright (c) 2011 AT&T Intellectual Property
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -12,4 +12,4 @@
 
 #include <cgraph/cgraph.h>
 
-    extern int run_gvpr(Agraph_t * srcGraph, int argc, char *argv[]);
+extern int run_gvpr(Agraph_t *srcGraph, size_t argc, char *argv[]);
